Splitting commands part 1, authentication
